refactor(audios): move fixtures to conftest.py
